Added in-place and scalar versions of xlogy. Added documentation comments for a number of APIs.
