+++ src/TorchSharp/Tensor/TorchTensor.Math.cs
@@ -15,6 +15,9 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_abs(IntPtr tensor);
 
+        /// <summary>
+        /// Compute the absolute value of each element in the tensor
+        /// </summary>
         public TorchTensor abs()
         {
             var res = THSTensor_abs(handle);
@@ -23,11 +26,17 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Compute the absolute value of each element in the tensor
+        /// </summary>
         public TorchTensor absolute() => abs();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_abs_(IntPtr tensor);
 
+        /// <summary>
+        /// Compute the absolute value of each element in the tensor, in-place.
+        /// </summary>
         public TorchTensor abs_()
         {
             var res = THSTensor_abs_(handle);
@@ -36,13 +45,16 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Compute the absolute value of each element in the tensor, in-place
+        /// </summary>
         public TorchTensor absolute_() => abs_();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_add(IntPtr tensor, IntPtr trg, IntPtr alpha);
 
         /// <summary>
-        /// 
+        /// Add two tensors, element-wise
         /// </summary>
         /// <param name="target"></param>
         /// <returns></returns>
@@ -52,7 +64,7 @@ namespace TorchSharp.Tensor
         }
 
         /// <summary>
-        /// 
+        /// Add two tensors, element-wise, scaling the second operator by 'alpha'
         /// </summary>
         /// <param name="target"></param>
         /// <param name="alpha"></param>
@@ -69,7 +81,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_add_scalar(IntPtr tensor, IntPtr trg, IntPtr alpha);
 
         /// <summary>
-        /// 
+        /// Add a scalar value to each element in the target tensor.
         /// </summary>
         /// <param name="scalar"></param>
         /// <returns></returns>
@@ -79,7 +91,7 @@ namespace TorchSharp.Tensor
         }
 
         /// <summary>
-        /// 
+        /// Add a scalar value to each element in the target tensor, scaled by 'alpha'
         /// </summary>
         /// <param name="scalar"></param>
         /// <param name="alpha"></param>
@@ -93,7 +105,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_add_(IntPtr tensor, IntPtr trg, IntPtr alpha);
 
         /// <summary>
-        /// 
+        /// In-place element-wise addition.
         /// </summary>
         /// <param name="target"></param>
         /// <returns></returns>
@@ -103,7 +115,7 @@ namespace TorchSharp.Tensor
         }
 
         /// <summary>
-        /// 
+        /// In-place element-wise addition, with scaling
         /// </summary>
         /// <param name="target"></param>
         /// <param name="alpha"></param>
@@ -117,7 +129,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_add_scalar_(IntPtr tensor, IntPtr trg, IntPtr alpha);
 
         /// <summary>
-        /// 
+        /// In-place scalar addition.
         /// </summary>
         /// <param name="scalar"></param>
         /// <returns></returns>
@@ -127,7 +139,7 @@ namespace TorchSharp.Tensor
         }
 
         /// <summary>
-        /// 
+        /// In-place scalar addition, scaled.
         /// </summary>
         /// <param name="scalar"></param>
         /// <param name="alpha"></param>
@@ -144,16 +156,18 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_addbmm(IntPtr mat, IntPtr batch1, IntPtr batch2, float beta, float alpha);
 
         /// <summary>
-        /// 
+        /// Performs a batch matrix-matrix product of matrices stored in batch1 and batch2, with a reduced
+        /// add step (all matrix multiplications get accumulated along the first dimension).
+        /// input is added to the final result.
         /// </summary>
-        /// <param name="batch1"></param>
-        /// <param name="batch2"></param>
+        /// <param name="mat1"></param>
+        /// <param name="mat2"></param>
         /// <param name="beta"></param>
         /// <param name="alpha"></param>
         /// <returns></returns>
-        public TorchTensor addbmm(TorchTensor batch1, TorchTensor batch2, float beta = 1, float alpha = 1)
+        public TorchTensor addbmm(TorchTensor mat1, TorchTensor mat2, float beta = 1, float alpha = 1)
         {
-            var res = THSTensor_addbmm(handle, batch1.Handle, batch2.Handle, beta, alpha);
+            var res = THSTensor_addbmm(handle, mat1.Handle, mat2.Handle, beta, alpha);
             if (res == IntPtr.Zero)
                 Torch.CheckForErrors();
             return new TorchTensor(res);
@@ -163,7 +177,9 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_addbmm_(IntPtr mat, IntPtr batch1, IntPtr batch2, float beta, float alpha);
 
         /// <summary>
-        /// 
+        /// Performs an in-place batch matrix-matrix product of matrices stored in batch1 and batch2, with a reduced
+        /// add step (all matrix multiplications get accumulated along the first dimension).
+        /// input is added to the final result.
         /// </summary>
         /// <param name="batch1"></param>
         /// <param name="batch2"></param>
@@ -182,7 +198,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_addcdiv(IntPtr tensor, IntPtr tensor1, IntPtr tensor2, IntPtr value);
 
         /// <summary>
-        /// 
+        /// Performs the element-wise division of tensor1 by tensor2, multiply the result by the scalar value and add it to input.
         /// </summary>
         /// <param name="tensor1"></param>
         /// <param name="tensor2"></param>
@@ -200,7 +216,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_addcdiv_(IntPtr tensor, IntPtr tensor1, IntPtr tensor2, IntPtr value);
 
         /// <summary>
-        /// 
+        /// Performs the in-place element-wise division of tensor1 by tensor2, multiply the result by the scalar value and add it to input.
         /// </summary>
         /// <param name="tensor1"></param>
         /// <param name="tensor2"></param>
@@ -218,7 +234,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_addcmul(IntPtr tensor, IntPtr tensor1, IntPtr tensor2, IntPtr value);
 
         /// <summary>
-        /// 
+        /// Performs the element-wise multiplication of tensor1 by tensor2, multiply the result by the scalar value and add it to input.
         /// </summary>
         /// <param name="tensor1"></param>
         /// <param name="tensor2"></param>
@@ -236,7 +252,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_addcmul_(IntPtr tensor, IntPtr tensor1, IntPtr tensor2, IntPtr value);
 
         /// <summary>
-        /// 
+        /// Performs the in-place element-wise multiplication of tensor1 by tensor2, multiply the result by the scalar value and add it to input.
         /// </summary>
         /// <param name="tensor1"></param>
         /// <param name="tensor2"></param>
@@ -255,7 +271,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_addmm(IntPtr mat, IntPtr mat1, IntPtr mat2, float beta, float alpha);
 
         /// <summary>
-        /// 
+        /// Performs a matrix multiplication of the matrices mat1 and mat2. The matrix input is added to the final result.
         /// </summary>
         /// <param name="mat1"></param>
         /// <param name="mat2"></param>
@@ -274,7 +290,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_addmm_(IntPtr mat, IntPtr mat1, IntPtr mat2, float beta, float alpha);
 
         /// <summary>
-        /// 
+        /// Performs an in-place matrix multiplication of the matrices mat1 and mat2. The matrix input is added to the final result.
         /// </summary>
         /// <param name="mat1"></param>
         /// <param name="mat2"></param>
@@ -370,6 +386,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_bitwise_and(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise bitwise AND
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor bitwise_and(TorchTensor other)
         {
             var res = THSTensor_bitwise_and(handle, other.Handle);
@@ -380,6 +401,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_bitwise_and_(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise bitwise AND, in-place
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor bitwise_and_(TorchTensor other)
         {
             var res = THSTensor_bitwise_and_(handle, other.Handle);
@@ -390,6 +416,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_bitwise_not(IntPtr tensor);
 
+        /// <summary>
+        /// Element-wise bitwise NOT
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor bitwise_not()
         {
             var res = THSTensor_bitwise_not(handle);
@@ -400,6 +430,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_bitwise_not_(IntPtr tensor);
 
+        /// <summary>
+        /// Element-wise bitwise NOT, in-place
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor bitwise_not_(TorchTensor other)
         {
             var res = THSTensor_bitwise_not_(handle);
@@ -410,6 +444,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_bitwise_or(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise bitwise OR
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor bitwise_or(TorchTensor other)
         {
             var res = THSTensor_bitwise_or(handle, other.Handle);
@@ -420,6 +458,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_bitwise_or_(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise bitwise OR, in-place
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor bitwise_or_(TorchTensor other)
         {
             var res = THSTensor_bitwise_or_(handle, other.Handle);
@@ -430,6 +472,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_bitwise_xor(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise bitwise XOR
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor bitwise_xor(TorchTensor other)
         {
             var res = THSTensor_bitwise_xor(handle, other.Handle);
@@ -440,6 +486,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_bitwise_xor_(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise bitwise XOR, in-place.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor bitwise_xor_(TorchTensor other)
         {
             var res = THSTensor_bitwise_xor_(handle, other.Handle);
@@ -450,6 +500,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_ceil(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the ceil of the elements of input, the smallest integer greater than or equal to each element.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor ceil()
         {
             var res = THSTensor_ceil(handle);
@@ -461,6 +515,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_ceil_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each element of the input with the smallest integer greater than or equal to the element.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor ceil_()
         {
             var res = THSTensor_ceil_(handle);
@@ -472,6 +530,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_div(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Divides each element of the input by the corresponding element of other.
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor div(TorchTensor target)
         {
             var res = THSTensor_div(handle, target.Handle);
@@ -479,22 +542,43 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Divides each element of the input by the corresponding element of other.
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor divide(TorchTensor target) => div(target);
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_div_scalar(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Scalar division
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor div(TorchScalar target)
         {
             var res = THSTensor_div_scalar(handle, target.Handle);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
             return new TorchTensor(res);
         }
+
+        /// <summary>
+        /// Scalar division
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor divide(TorchScalar target) => div(target);
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_div_(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// In-place division
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor div_(TorchTensor target)
         {
             var res = THSTensor_div_(handle, target.Handle);
@@ -505,21 +589,15 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_div_scalar_(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// In-place scalar division
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor div_(TorchScalar target)
         {
             var res = THSTensor_div_scalar_(handle, target.Handle);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_floor(IntPtr tensor);
-
-        public TorchTensor floor()
-        {
-            var res = THSTensor_floor(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
             return new TorchTensor(res);
         }
 
@@ -536,6 +614,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_exp_(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the exponential of the elements of the input tensor input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor exp_()
         {
             var res = THSTensor_exp_(handle);
@@ -546,6 +628,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_exp2(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the base 2 exponential function of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor exp2()
         {
             var res = THSTensor_exp2(handle);
@@ -556,6 +642,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_expm1(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the exponential of the elements minus 1 of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor expm1()
         {
             var res = THSTensor_expm1(handle);
@@ -565,6 +655,11 @@ namespace TorchSharp.Tensor
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_expm1_(IntPtr tensor);
+
+        /// <summary>
+        /// Replaces each element with the exponential of the element minus 1 of input.
+        /// </summary>
+        /// <returns></returns>
 
         public TorchTensor expm1_()
         {
@@ -616,6 +711,12 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_float_power(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Raises input to the power of exponent, elementwise, in double precision.
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
+        /// <remarks> If neither input is complex returns a torch.float64 tensor, and if one or more inputs is complex returns a torch.complex128 tensor.</remarks>
         public TorchTensor float_power(TorchTensor target)
         {
             var res = THSTensor_float_power(handle, target.Handle);
@@ -624,8 +725,27 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
+        static extern IntPtr THSTensor_floor(IntPtr tensor);
+
+        /// <summary>
+        /// Returns a new tensor with the floor of the elements of input, the largest integer less than or equal to each element.
+        /// </summary>
+        /// <returns></returns>
+        public TorchTensor floor()
+        {
+            var res = THSTensor_floor(handle);
+            if (res == IntPtr.Zero)
+                Torch.CheckForErrors();
+            return new TorchTensor(res);
+        }
+
+        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_floor_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each element with the floor of the input, the largest integer less than or equal to each element.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor floor_()
         {
             var res = THSTensor_floor_(handle);
@@ -637,6 +757,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_fmod(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Computes the element-wise remainder of division.
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor fmod(TorchTensor target)
         {
             var res = THSTensor_fmod(handle, target.Handle);
@@ -647,6 +772,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_fmod_(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Computes the element-wise remainder of division, in-place.
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor fmod_(TorchTensor target)
         {
             var res = THSTensor_fmod_(handle, target.Handle);
@@ -657,6 +787,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_fmod_scalar(IntPtr tensor, IntPtr scalar);
 
+        /// <summary>
+        /// Computes the element-wise remainder of division.
+        /// </summary>
+        /// <param name="scalar"></param>
+        /// <returns></returns>
         public TorchTensor fmod(TorchScalar scalar)
         {
             var res = THSTensor_fmod_scalar(handle, scalar.Handle);
@@ -667,6 +802,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_fmod_scalar_(IntPtr tensor, IntPtr scalar);
 
+        /// <summary>
+        /// Computes the element-wise remainder of division, in-place
+        /// </summary>
+        /// <param name="scalar"></param>
+        /// <returns></returns>
         public TorchTensor fmod_(TorchScalar scalar)
         {
             var res = THSTensor_fmod_scalar_(handle, scalar.Handle);
@@ -677,6 +817,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_frac(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the fractional portion of each element in input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor frac()
         {
             var res = THSTensor_frac(handle);
@@ -687,6 +831,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_frac_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the fractional portion of each element in input, in-place.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor frac_()
         {
             var res = THSTensor_frac_(handle);
@@ -707,6 +855,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_gcd_(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Computes the element-wise greatest common divisor (GCD) of input and other.
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor gcd_(TorchTensor other)
         {
             var res = THSTensor_gcd_(handle, other.Handle);
@@ -717,6 +870,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_hypot(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise: given the legs of a right triangle, return its hypotenuse.
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor hypot(TorchTensor other)
         {
             var res = THSTensor_hypot(handle, other.handle);
@@ -725,10 +883,13 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
-
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the natural logarithm of the input elements.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor log()
         {
             var res = THSTensor_log(handle);
@@ -739,6 +900,9 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each elements with the natural logarithm of the input.
+        /// </summary>
         public TorchTensor log_()
         {
             var res = THSTensor_log_(handle);
@@ -749,6 +913,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logaddexp(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Logarithm of the sum of exponentiations of the inputs.
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logaddexp(TorchTensor other)
         {
             var res = THSTensor_logaddexp(handle, other.handle);
@@ -760,6 +929,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logaddexp2(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Logarithm of the sum of exponentiations of the inputs in base-2.
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logaddexp2(TorchTensor other)
         {
             var res = THSTensor_logaddexp2(handle, other.handle);
@@ -771,6 +945,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logcumsumexp(IntPtr tensor, long dim);
 
+        /// <summary>
+        /// Returns the logarithm of the cumulative summation of the exponentiation of elements of input in the dimension dim.
+        /// </summary>
+        /// <param name="dim"></param>
+        /// <returns></returns>
         public TorchTensor logcumsumexp(long dim)
         {
             var res = THSTensor_logcumsumexp(handle, dim);
@@ -782,6 +961,13 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logsumexp(IntPtr tensor, long dim, bool keepdim);
 
+        /// <summary>
+        /// Returns the log of summed exponentials of each row of the input tensor in the given dimension dim. 
+        /// </summary>
+        /// <param name="dim"></param>
+        /// <param name="keepdim"></param>
+        /// <returns></returns>
+        /// <remarks>The computation is numerically stabilized.</remarks>
         public TorchTensor logsumexp(long dim, Boolean keepdim = false)
         {
             var res = THSTensor_logsumexp(handle, dim, keepdim);
@@ -793,6 +979,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log10(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the logarithm to the base 10 of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor log10()
         {
             var res = THSTensor_log10(handle);
@@ -804,6 +994,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log10_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces elements with the logarithm to the base 10 of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor log10_()
         {
             var res = THSTensor_log10_(handle);
@@ -815,6 +1009,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log1p(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the natural logarithm of (1 + input).
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor log1p()
         {
             var res = THSTensor_log1p(handle);
@@ -826,6 +1024,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log1p_(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the natural logarithm of (1 + input), inplace.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor log1p_()
         {
             var res = THSTensor_log1p_(handle);
@@ -836,6 +1038,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log2(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the logarithm to the base 2 of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor log2()
         {
             var res = THSTensor_log2(handle);
@@ -847,6 +1053,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log2_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each element with the logarithm to the base 2 of the input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor log2_()
         {
             var res = THSTensor_log2_(handle);
@@ -857,6 +1067,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logical_and(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Logical AND
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logical_and(TorchTensor other)
         {
             var res = THSTensor_logical_and(handle, other.Handle);
@@ -867,6 +1082,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logical_and_(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Logical AND, in place
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logical_and_(TorchTensor other)
         {
             var res = THSTensor_logical_and_(handle, other.Handle);
@@ -877,6 +1097,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logical_not(IntPtr tensor);
 
+        /// <summary>
+        /// Logical NOT
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor logical_not()
         {
             var res = THSTensor_logical_not(handle);
@@ -887,6 +1111,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logical_not_(IntPtr tensor);
 
+        /// <summary>
+        /// Logical NOT, in place
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logical_not_(TorchTensor other)
         {
             var res = THSTensor_logical_not_(handle);
@@ -897,6 +1126,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logical_or(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Logical OR
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logical_or(TorchTensor other)
         {
             var res = THSTensor_logical_or(handle, other.Handle);
@@ -907,6 +1141,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logical_or_(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Logical OR, in place
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logical_or_(TorchTensor other)
         {
             var res = THSTensor_logical_or_(handle, other.Handle);
@@ -917,6 +1156,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logical_xor(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Logical XOR
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logical_xor(TorchTensor other)
         {
             var res = THSTensor_logical_xor(handle, other.Handle);
@@ -927,6 +1171,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logical_xor_(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Logical XOR, in place
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor logical_xor_(TorchTensor other)
         {
             var res = THSTensor_logical_xor_(handle, other.Handle);
@@ -937,6 +1186,12 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_logit(IntPtr tensor, IntPtr eps);
 
+        /// <summary>
+        /// Returns a new tensor with the logit of the elements of input.
+        /// input is clamped to [eps, 1 - eps] when eps is not null
+        /// </summary>
+        /// <param name="eps"></param>
+        /// <returns></returns>
         public TorchTensor logit(double? eps = null)
         {
             var epsArr = eps.HasValue ? new double[] { eps.Value } : null;
@@ -953,6 +1208,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_mul(IntPtr tensor, IntPtr target);
 
+        /// <summary>
+        /// Element-wise multiplication
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor mul(TorchTensor target)
         {
             var res = THSTensor_mul(handle, target.Handle);
@@ -960,23 +1220,43 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Element-wise multiplication
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor multiply(TorchTensor target) => mul(target);
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_mul_scalar(IntPtr tensor, IntPtr scalar);
 
-        public TorchTensor mul(TorchScalar scalar)
+        /// <summary>
+        /// Element-wise multiplication
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
+        public TorchTensor mul(TorchScalar target)
         {
-            var res = THSTensor_mul_scalar(handle, scalar.Handle);
+            var res = THSTensor_mul_scalar(handle, target.Handle);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Element-wise multiplcation
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor multiply(TorchScalar target) => mul(target);
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_mul_(IntPtr tensor, IntPtr target);
 
+        /// <summary>
+        /// Element-wise multiplication, in place
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor mul_(TorchTensor target)
         {
             var res = THSTensor_mul_(handle, target.Handle);
@@ -987,6 +1267,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_mul_scalar_(IntPtr tensor, IntPtr target);
 
+        /// <summary>
+        /// Element-wise multiplication, in place
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor mul_(TorchScalar target)
         {
             var res = THSTensor_mul_scalar_(handle, target.Handle);
@@ -1002,6 +1287,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_neg(IntPtr tensor);
 
+        /// <summary>
+        /// Negation
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor neg()
         {
             var res = THSTensor_neg(handle);
@@ -1009,11 +1298,19 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Negation
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor negative() => neg();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_neg_(IntPtr tensor);
 
+        /// <summary>
+        /// In-place negation
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor neg_()
         {
             var res = THSTensor_neg_(handle);
@@ -1024,6 +1321,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_pow(IntPtr tensor, IntPtr exponent);
 
+        /// <summary>
+        /// Takes the power of each element in input with exponent and returns a tensor with the result.
+        /// </summary>
+        /// <param name="exponent"></param>
+        /// <returns></returns>
         public TorchTensor pow(TorchTensor exponent)
         {
             var res = THSTensor_pow(handle, exponent.Handle);
@@ -1034,6 +1336,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_pow_(IntPtr tensor, IntPtr exponent);
 
+        /// <summary>
+        /// Replaces each element in input with the power of the element and the exponent.
+        /// </summary>
+        /// <param name="exponent"></param>
+        /// <returns></returns>
         public TorchTensor pow_(TorchTensor exponent)
         {
             var res = THSTensor_pow_(handle, exponent.Handle);
@@ -1044,9 +1351,14 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_pow_scalar(IntPtr tensor, IntPtr scalar);
 
-        public TorchTensor pow(TorchScalar scalar)
+        /// <summary>
+        /// Takes the power of each element in input with exponent and returns a tensor with the result.
+        /// </summary>
+        /// <param name="exponent"></param>
+        /// <returns></returns>
+        public TorchTensor pow(TorchScalar exponent)
         {
-            var res = THSTensor_pow_scalar(handle, scalar.Handle);
+            var res = THSTensor_pow_scalar(handle, exponent.Handle);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
             return new TorchTensor(res);
         }
@@ -1054,9 +1366,14 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_pow_scalar_(IntPtr tensor, IntPtr scalar);
 
-        public TorchTensor pow_(TorchScalar scalar)
+        /// <summary>
+        /// Replaces each element in input with the power of the element and the exponent.
+        /// </summary>
+        /// <param name="exponent"></param>
+        /// <returns></returns>
+        public TorchTensor pow_(TorchScalar exponent)
         {
-            var res = THSTensor_pow_scalar_(handle, scalar.Handle);
+            var res = THSTensor_pow_scalar_(handle, exponent.Handle);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
             return new TorchTensor(res);
         }
@@ -1064,6 +1381,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_reciprocal(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the reciprocal of the elements of input
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor reciprocal()
         {
             var res = THSTensor_reciprocal(handle);
@@ -1075,6 +1396,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_reciprocal_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each element with the reciprocal of the input
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor reciprocal_()
         {
             var res = THSTensor_reciprocal_(handle);
@@ -1086,6 +1411,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_remainder(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Computes the element-wise remainder of division.
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor remainder(TorchTensor target)
         {
             var res = THSTensor_remainder(handle, target.Handle);
@@ -1096,6 +1426,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_remainder_(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Computes the element-wise remainder of division, in place
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor remainder_(TorchTensor target)
         {
             var res = THSTensor_remainder_(handle, target.Handle);
@@ -1106,6 +1441,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_remainder_scalar(IntPtr tensor, IntPtr scalar);
 
+        /// <summary>
+        /// Computes the element-wise remainder of division.
+        /// </summary>
+        /// <param name="scalar"></param>
+        /// <returns></returns>
         public TorchTensor remainder(TorchScalar scalar)
         {
             var res = THSTensor_remainder_scalar(handle, scalar.Handle);
@@ -1117,6 +1457,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_remainder_scalar_(IntPtr tensor, IntPtr scalar);
 
+        /// <summary>
+        /// Computes the element-wise remainder of division.
+        /// </summary>
+        /// <param name="scalar"></param>
+        /// <returns></returns>
         public TorchTensor remainder_(TorchScalar scalar)
         {
             var res = THSTensor_remainder_scalar_(handle, scalar.Handle);
@@ -1127,6 +1472,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_round(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with each of the elements of input rounded to the closest integer.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor round()
         {
             var res = THSTensor_round(handle);
@@ -1138,6 +1487,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_round_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each of the elements of input with the element rounded to the closest integer.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor round_()
         {
             var res = THSTensor_round_(handle);
@@ -1149,6 +1502,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_rsqrt(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the reciprocal of the square-root of each of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor rsqrt()
         {
             var res = THSTensor_rsqrt(handle);
@@ -1159,6 +1516,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_rsqrt_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each of the elements of input with  the reciprocal of the square-root of each of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor rsqrt_()
         {
             var res = THSTensor_rsqrt_(handle);
@@ -1169,6 +1530,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sqrt(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the element-wise square root
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor sqrt()
         {
             var res = THSTensor_sqrt(handle);
@@ -1179,6 +1544,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sqrt_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the element-wise square root, in place
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor sqrt_()
         {
             var res = THSTensor_sqrt_(handle);
@@ -1189,6 +1558,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sign(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the signs (-1, 0, 1) of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor sign()
         {
             var res = THSTensor_sign(handle);
@@ -1200,6 +1573,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sign_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each element with the signs (-1, 0, 1) of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor sign_()
         {
             var res = THSTensor_sign_(handle);
@@ -1211,6 +1588,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_signbit(IntPtr tensor);
 
+        /// <summary>
+        /// Tests if each element of input has its sign bit set (is less than zero) or not.
+        /// </summary>
+        /// <returns>A boolean tensor of the same shape as the input.</returns>
         public TorchTensor signbit()
         {
             var res = THSTensor_signbit(handle);
@@ -1222,6 +1603,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sub(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Element-wise subtraction
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor sub(TorchTensor target)
         {
             var res = THSTensor_sub(handle, target.Handle);
@@ -1232,6 +1618,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sub_scalar(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Element-wise subtraction
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor sub(TorchScalar target)
         {
             var res = THSTensor_sub_scalar(handle, target.Handle);
@@ -1242,6 +1633,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sub_(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Element-wise subtraction, in place
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor sub_(TorchTensor target)
         {
             var res = THSTensor_sub_(handle, target.Handle);
@@ -1252,6 +1648,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sub_scalar_(IntPtr tensor, IntPtr trg);
 
+        /// <summary>
+        /// Element-wise subtraction, in-place
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
         public TorchTensor sub_(TorchScalar target)
         {
             var res = THSTensor_sub_scalar_(handle, target.Handle);
@@ -1262,6 +1663,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_trunc(IntPtr tensor);
 
+        /// <summary>
+        /// Returns a new tensor with the truncated integer values of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor trunc()
         {
             var res = THSTensor_trunc(handle);
@@ -1269,11 +1674,19 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Returns a new tensor with the truncated integer values of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor fix() => trunc();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_trunc_(IntPtr tensor);
 
+        /// <summary>
+        /// Replaces each element with the truncated integer values of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor trunc_()
         {
             var res = THSTensor_trunc_(handle);
@@ -1281,6 +1694,10 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Replaces each element with the truncated integer values of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor fix_() => trunc_();
 
         [DllImport("LibTorchSharp")]
@@ -1289,11 +1706,61 @@ namespace TorchSharp.Tensor
         /// <summary>
         /// Computes x * log(y)
         /// </summary>
-        /// <param name="target"></param>
+        /// <param name="y"></param>
         /// <returns></returns>
-        public TorchTensor xlogy(TorchTensor target)
+        public TorchTensor xlogy(TorchTensor y)
         {
-            var res = THSTensor_xlogy(handle, target.Handle);
+            var res = THSTensor_xlogy(handle, y.Handle);
+            if (res == IntPtr.Zero)
+                Torch.CheckForErrors();
+            return new TorchTensor(res);
+        }
+
+
+        [DllImport("LibTorchSharp")]
+        static extern IntPtr THSTensor_xlogy_(IntPtr tensor, IntPtr trg);
+
+        /// <summary>
+        /// Computes x * log(y) in place
+        /// </summary>
+        /// <param name="y"></param>
+        /// <returns></returns>
+        public TorchTensor xlogy_(TorchTensor y)
+        {
+            var res = THSTensor_xlogy_(handle, y.Handle);
+            if (res == IntPtr.Zero)
+                Torch.CheckForErrors();
+            return new TorchTensor(res);
+        }
+
+        [DllImport("LibTorchSharp")]
+        static extern IntPtr THSTensor_xlogy_scalar(IntPtr tensor, IntPtr trg);
+
+        /// <summary>
+        /// Computes x * log(y)
+        /// </summary>
+        /// <param name="y"></param>
+        /// <returns></returns>
+        public TorchTensor xlogy(TorchScalar y)
+        {
+            var res = THSTensor_xlogy_scalar(handle, y.Handle);
+            if (res == IntPtr.Zero)
+                Torch.CheckForErrors();
+            return new TorchTensor(res);
+        }
+
+
+        [DllImport("LibTorchSharp")]
+        static extern IntPtr THSTensor_xlogy_scalar_(IntPtr tensor, IntPtr trg);
+
+        /// <summary>
+        /// Computes x * log(y) in place
+        /// </summary>
+        /// <param name="y"></param>
+        /// <returns></returns>
+        public TorchTensor xlogy_(TorchScalar y)
+        {
+            var res = THSTensor_xlogy_scalar_(handle, y.Handle);
             if (res == IntPtr.Zero)
                 Torch.CheckForErrors();
             return new TorchTensor(res);
