+++ src/TorchSharp/Tensor/TorchTensor.Trig.cs
@@ -7,15 +7,22 @@ using System.Runtime.InteropServices;
 using System.Text;
 namespace TorchSharp.Tensor
 {
+    // This files contains trigonometric and hyperbolic functions.
+    // All are element-wise operations on tensors.
+
     public sealed partial class TorchTensor
     {
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_angle(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the element-wise angle (in radians) of the given input tensor.
         /// </summary>
         /// <returns></returns>
+        /// <remarks>
+        /// Starting in Torch 1.8, angle returns pi for negative real numbers, zero for non-negative real numbers, and propagates NaNs.
+        /// Previously the function would return zero for all real numbers and not propagate floating-point NaNs.
+        /// </remarks>
         public TorchTensor angle()
         {
             return new TorchTensor(THSTensor_angle(handle));
@@ -25,7 +32,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_asin(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the arcsine of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor asin()
@@ -33,13 +40,17 @@ namespace TorchSharp.Tensor
             return new TorchTensor(THSTensor_asin(handle));
         }
 
+        /// <summary>
+        /// Computes the arcsine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arcsin() => asin();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_asin_(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the arcsine of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor asin_()
@@ -56,7 +67,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_acos(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the arccosine of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor acos()
@@ -67,13 +78,17 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the arccosine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arccos() => acos();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_acos_(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the arccosine of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor acos_()
@@ -84,13 +99,17 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the arccosine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arccos_() => acos_();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_atan(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the arctangent of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor atan()
@@ -101,11 +120,19 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the arctangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arctan() => atan();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_atan_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the arctangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor atan_()
         {
             var res = THSTensor_atan_(handle);
@@ -114,11 +141,20 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the arctangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arctan_() => atan_();
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_atan2(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise arctangent of input / other with consideration of the quadrant.
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor atan2(TorchTensor other)
         {
             var res = THSTensor_atan2(handle, other.Handle);
@@ -130,6 +166,11 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_atan2_(IntPtr tensor, IntPtr other);
 
+        /// <summary>
+        /// Element-wise arctangent of input / other with consideration of the quadrant.
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
         public TorchTensor atan2_(TorchTensor other)
         {
             var res = THSTensor_atan2_(handle, other.Handle);
@@ -142,7 +183,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_cos(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the cosine of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor cos()
@@ -157,7 +198,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_cos_(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the cosine of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor cos_()
@@ -172,7 +213,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_sin(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the sine of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor sin()
@@ -187,7 +228,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_sin_(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the sine of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor sin_()
@@ -202,7 +243,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_tan(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the tangent of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor tan()
@@ -217,7 +258,7 @@ namespace TorchSharp.Tensor
         static extern IntPtr THSTensor_tan_(IntPtr tensor);
 
         /// <summary>
-        /// 
+        /// Computes the tangent of the elements of input.
         /// </summary>
         /// <returns></returns>
         public TorchTensor tan_()
@@ -231,6 +272,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sinc(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the normalized sinc of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor sinc()
         {
             var res = THSTensor_sinc(handle);
@@ -242,6 +287,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sinc_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the normalized sinc of input, in place.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor sinc_()
         {
             var res = THSTensor_sinc_(handle);
@@ -253,6 +302,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sinh(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic sine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor sinh()
         {
             var res = THSTensor_sinh(handle);
@@ -264,6 +317,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_sinh_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic sine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor sinh_()
         {
             var res = THSTensor_sinh_(handle);
@@ -275,6 +332,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_cosh(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic cosine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor cosh()
         {
             var res = THSTensor_cosh(handle);
@@ -286,6 +347,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_cosh_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic cosine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor cosh_()
         {
             var res = THSTensor_cosh_(handle);
@@ -296,6 +361,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_tanh(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic tangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor tanh()
         {
             var res = THSTensor_tanh(handle);
@@ -307,6 +376,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_tanh_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic tangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor tanh_()
         {
             var res = THSTensor_tanh_(handle);
@@ -318,6 +391,10 @@ namespace TorchSharp.Tensor
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_arcsinh(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic arcsine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arcsinh()
         {
             var res = THSTensor_arcsinh(handle);
@@ -326,9 +403,19 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the hyperbolic arcsine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
+        public TorchTensor asinh() => arcsinh();
+
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_arcsinh_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic arcsine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arcsinh_()
         {
             var res = THSTensor_arcsinh_(handle);
@@ -337,9 +424,19 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the hyperbolic arcsine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
+        public TorchTensor asinh_() => arcsinh_();
+
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_arccosh(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic arccosine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arccosh()
         {
             var res = THSTensor_arccosh(handle);
@@ -348,9 +445,20 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the hyperbolic arccosine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
+        public TorchTensor acosh() => arccosh();
+
+
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_arccosh_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic arccosine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arccosh_()
         {
             var res = THSTensor_arccosh_(handle);
@@ -359,9 +467,19 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the hyperbolic arccosine of the elements of input.
+        /// </summary>
+        /// <returns></returns>
+        public TorchTensor acosh_() => arccosh_();
+
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_arctanh(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic arctangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arctanh()
         {
             var res = THSTensor_arctanh(handle);
@@ -370,9 +488,19 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
+        /// <summary>
+        /// Computes the hyperbolic arctangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
+        public TorchTensor atanh() => arctanh();
+
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_arctanh_(IntPtr tensor);
 
+        /// <summary>
+        /// Computes the hyperbolic arctangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor arctanh_()
         {
             var res = THSTensor_arctanh_(handle);
@@ -381,17 +509,10 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
-        public TorchTensor asinh() => arcsinh();
-
-        public TorchTensor asinh_() => arctanh_();
-
-        public TorchTensor acosh() => arccosh();
-
-        public TorchTensor acosh_() => arccosh_();
-
-        public TorchTensor atanh() => arctanh();
-
+        /// <summary>
+        /// Computes the hyperbolic arctangent of the elements of input.
+        /// </summary>
+        /// <returns></returns>
         public TorchTensor atanh_() => arctanh_();
-
     }
 }
